Reorganized project structure, moving all views into a Views folder
