+++ ios/arGPT/arGPT/Views/ColorModeAdaptiveImage.swift
